refactor(ui-grid): extract useGridScrollSync
